+++ src/main/java/commands/EnumCommand.java
@@ -1,3 +1,5 @@
+package commands;
+
 public enum EnumCommand {
     DEADLINE, EVENT, TODO, BYE, DONE, DELETE, LIST, CHECK;
 }
