+++ src/main/java/tasks/Deadlines.java
@@ -1,3 +1,7 @@
+package tasks;
+
+import tasks.Task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
