+++ text-ui-test/EXPECTED.TXT
@@ -55,13 +55,6 @@
     ___________________________________________________________________________________
 
     ___________________________________________________________________________________
-     :) Here are all the tasks in your list:
-     1. [T][✓] readbook
-     2. [T][✘] join sports club
-     3. [T][✘] borrow book
-    ___________________________________________________________________________________
-
-    ___________________________________________________________________________________
      ☹ OOPS!!! The description of a deadline cannot be empty.
     ___________________________________________________________________________________
